Added Island Generator v2.0 + Working Island Generator + Working Island Deleter + Working Generator when Empty Island is available + Added WorldEdit and WorldGuard Dependencies( Paste schamtics is avaliable ) + Corrected HelpCommand( variables sender etc. ) + Added IslandCommand with args[ create, delete ], definition in Main and plugin.yml usage + Added Execute Function onEnable + Created pre-yml [freeislands] + Corrected [islands.yml, nextisland.yml] ! I don't know changes in Islands.java xD
+ Teleport to island on create
+ Deleted inline comments like "Lucyfer zjebau"
================================================================
+ Added foreach loop in loadIsland()
+ Added isPlayerOwner boolean
+ Added isPlayerMember boolean
+ Added spawn argument in Class Island

+++ src/main/java/pl/trollcraft/Skyblock/commands/IslandCommand.java
@@ -8,6 +8,7 @@ import org.jetbrains.annotations.NotNull;
 import pl.trollcraft.Skyblock.essentials.ChatUtils;
 import pl.trollcraft.Skyblock.generator.CreateIsland;
 import pl.trollcraft.Skyblock.generator.DeleteIsland;
+import pl.trollcraft.Skyblock.island.Islands;
 
 public class IslandCommand implements CommandExecutor {
 
@@ -23,12 +24,27 @@ public class IslandCommand implements CommandExecutor {
         }
         else{
             if( args[0].equalsIgnoreCase("create")){
-                sender.sendMessage(ChatUtils.fixColor("&aTworze wyspe..."));
-                CreateIsland.createNew(player);
+                if(Islands.isPlayerOwner(player.getName())){
+                    sender.sendMessage(ChatUtils.fixColor("&cPosiadasz juz wyspe!"));
+                }
+                else {
+                    sender.sendMessage(ChatUtils.fixColor("&aTworze wyspe..."));
+                    CreateIsland.createNew(player);
+                }
             }
             else if( args[0].equalsIgnoreCase("delete")){
-                DeleteIsland.deleteIs(player.getName());
-                sender.sendMessage("Usunieto wyspe");
+                if(Islands.isPlayerOwner(player.getName())) {
+                    DeleteIsland.deleteIs(player.getName());
+                    sender.sendMessage(ChatUtils.fixColor("&aUsunieto wyspe"));
+                }
+                else{
+                    sender.sendMessage(ChatUtils.fixColor("&cNie posiadasz wyspy"));
+                }
+            }
+            else if( args[0].equalsIgnoreCase("add")){
+                if(Islands.isPlayerOwner(player.getName())) {
+                    Islands.getIslands();
+                }
             }
         }
         return true;
